Refactor Text component files. Export as experimental component in src/index
