--- packages/components/src/text/font-family.native.js
@@ -1,1 +1,0 @@
-export const fontFamily = '';
