+++ packages/components/src/text/index.native.js
@@ -1,13 +1,13 @@
 /**
  * External dependencies
  */
-import styled from '@emotion/styled';
+import styled from '@emotion/native';
 
 /**
  * Internal dependencies
  */
 import { text } from './styles/text-mixins';
 
-const Text = styled.p( `margin: 0;`, text );
+const Text = styled.Text( text );
 
 export default Text;
