+++ packages/components/src/text/styles/text-styles.js
@@ -6,6 +6,6 @@ import styled from '@emotion/styled';
 /**
  * Internal dependencies
  */
-import { text } from './mixins';
+import { text } from './text-mixins';
 
 export const __experimentalText = styled.p( `margin: 0;`, text );
