+++ packages/components/src/text/styles/text-styles.native.js
@@ -6,6 +6,6 @@ import styled from '@emotion/native';
 /**
  * Internal dependencies
  */
-import { text } from './mixins';
+import { text } from './text-mixins';
 
 export const __experimentalText = styled.Text( text );
